Add necessary packages for vignettes
This packages are used in vignettes, but they were not being declared in
the DESCRIPTION file. This was causing NOTEs in checks. The solution is
to put theses packages in the `Suggests` field, as stated in R
Extensions Manual:

The ‘Suggests’ field uses the same syntax as ‘Depends’ and lists
packages that are not necessarily needed. This includes packages used
only in examples, tests or vignettes (see Section 1.4 [Writing package
vignettes], page 26), and packages loaded in the body of functions.
E.g., suppose an example from package foo uses a dataset from package
bar. Then it is not necessary to have bar use foo unless one wants to
execute all the examples/tests/vignettes: it is useful to have bar, but
not necessary. Version requirements can be specified, and will be used
by R CMD check.